+++ packages/node/CHANGELOG.md
@@ -14,6 +14,29 @@
   - generaltranslation@9.0.0-odysseus.0
   - gt-i18n@1.0.0-odysseus.0
 
+## 0.7.8
+
+### Patch Changes
+
+- Updated dependencies []:
+  - gt-i18n@0.9.8
+
+## 0.7.7
+
+### Patch Changes
+
+- Updated dependencies [[`9709a2f`](https://github.com/generaltranslation/gt/commit/9709a2f2b97b9d8239298e39bb31e57692bbffd8)]:
+  - generaltranslation@8.2.18
+  - gt-i18n@0.9.7
+
+## 0.7.6
+
+### Patch Changes
+
+- Updated dependencies [[`3197028`](https://github.com/generaltranslation/gt/commit/319702855a7b129f95217d41be9f2402680a2f01)]:
+  - generaltranslation@8.2.17
+  - gt-i18n@0.9.6
+
 ## 0.7.5
 
 ### Patch Changes